requirements.txt
beautifulsoup4>=4.11.1
fitz>=0.0.1.dev2
frontend>=0.0.3
pandas>=1.5.0
Pillow>=9.2.0
psycopg2
PyMuPDF>=1.20.2
PyPDF2>=2.11.0
pytesseract>=0.3.10
sqlalchemy>=1.4.41
requests-html>=0.10.0
duckpy>=3.2.0